Ajout de la classe Attack, tentative d'ajout des observeurs dans la classe deplacement et Loop (LoopEnemi), debut du code behind

+++ src/model/entities/Props.java
@@ -1,11 +1,9 @@
 package model.entities;
 
-import model.entities.Entite;
-
 public class Props extends Entite {
 
-    public Props(int xSize, int ySize, int xpos, int ypos, String id) {
+    public Props(double xSize, double ySize, double xpos, double ypos, String id) {
 
-        super(xSize, ySize, xpos, ypos, id);
+        super((int)xSize, (int)ySize, (int)xpos, (int)ypos, id);
     }
 }
